refactor(utils): replace old utils file with experimental
