+++ shikithon/resources/user_images.py
@@ -4,7 +4,6 @@ from typing import Any, Dict, Optional, Union
 from ..decorators import method_endpoint
 from ..enums import RequestType
 from ..models import CreatedUserImage
-from ..utils import ExperimentalUtils
 from ..utils import Utils
 from .base_resource import BaseResource
 
@@ -31,10 +30,10 @@ class UserImages(BaseResource):
         :return: Created image info
         :rtype: Optional[CreatedUserImage]
         """
-        image_data = await ExperimentalUtils.get_image_data(image_path)
+        image_data = await Utils.get_image_data(image_path)
         response: Union[Dict[str, Any], int] = await self._client.request(
             self._client.endpoints.user_images,
-            data=ExperimentalUtils.create_data_dict(linked_type=linked_type),
+            data=Utils.create_data_dict(linked_type=linked_type),
             bytes_data=image_data,
             request_type=RequestType.POST)
         return Utils.validate_response_data(response,
