+++ shikithon/utils/__init__.py
@@ -1,6 +1,5 @@
 """Utils for shikithon library."""
 
-from .experimental_utils import ExperimentalUtils
 from .utils import Utils
 
-__all__ = ['Utils', 'ExperimentalUtils']
+__all__ = ['Utils']
